transform_camera_work
need to optimize
retired non use, non util functions
more maps ?

+++ header/mini_rt.h
@@ -44,7 +44,7 @@ void			add_maps(t_maps **head, t_maps *new_elem);
 
 void			ft_translation(t_vect *point, int move, double incr);
 void			ft_scale(double *value, int move, double incr);
-void			ft_rotate(t_vect *direction, int keycode, int incr_deg);
+void			ft_rotate(t_vect *direction, int keycode, double incr_deg);
 
 int				select_primary(t_scene *scene, int select);
 t_obj			is_primary_selected(t_scene *scene);
